Move BinarySclar class from expression.scalar.arithmetic package to expression.scalar
