+++ server/src/main/java/io/crate/expression/scalar/bitwise/BitwiseFunctions.java
@@ -23,7 +23,7 @@ package io.crate.expression.scalar.bitwise;
 
 import io.crate.common.TriConsumer;
 import io.crate.expression.scalar.ScalarFunctionModule;
-import io.crate.expression.scalar.arithmetic.BinaryScalar;
+import io.crate.expression.scalar.BinaryScalar;
 import io.crate.expression.scalar.Scalar;
 import io.crate.metadata.functions.Signature;
 import io.crate.sql.tree.BitString;
