+++ server/src/main/java/io/crate/expression/scalar/string/EncodeDecodeFunction.java
@@ -30,7 +30,7 @@ import java.util.function.UnaryOperator;
 import io.crate.common.Hex;
 import io.crate.common.Octal;
 import io.crate.expression.scalar.ScalarFunctionModule;
-import io.crate.expression.scalar.arithmetic.BinaryScalar;
+import io.crate.expression.scalar.BinaryScalar;
 import io.crate.expression.scalar.Scalar;
 import io.crate.metadata.functions.Signature;
 import io.crate.types.DataTypes;
