+++ server/src/main/java/io/crate/expression/scalar/systeminformation/PgGetSerialSequenceFunction.java
@@ -22,7 +22,7 @@
 package io.crate.expression.scalar.systeminformation;
 
 import io.crate.expression.scalar.ScalarFunctionModule;
-import io.crate.expression.scalar.arithmetic.BinaryScalar;
+import io.crate.expression.scalar.BinaryScalar;
 import io.crate.metadata.FunctionName;
 import io.crate.expression.scalar.Scalar;
 import io.crate.metadata.functions.Signature;
